requirements.txt
pycairo==1.21.0
cairosvg==2.5.2